Refactor supraland_parser.py to split into multiple modules .gitignore __pycache__ directory
