+++ scripts/parserlib/fileio.py
@@ -1,0 +1,151 @@
+import json
+from pathlib import Path
+import sys
+from typing import Any, Optional
+from .utils import JsonData
+
+
+# The following load/save functions make a path by joining all the varargs
+# together with '\\' and adding '.txt'
+#
+# For game specific source data this is usually something like:
+#   sourcedir:  ..\\source             (but can be any path)
+#   game:       'sl', 'siu', 'sw'      (note Crash files are in sl source directories)
+#   dir:        'levels', 'enums', 'bps', 'loc', 'mapimg'
+#   filename:   'Map'
+#
+# For output data used by the web map it's likely:
+#  '..\\public\\data', 'gameClasses'
+#
+# Extension should not be included. Empty strings will be ignored/skipped
+
+
+# Reads a json file and converts it do an object data structure
+def load_json_file(path: Path, quiet: bool = False):
+    if not path.exists():
+        sys.exit(f'{path} not found, exiting')
+    if not quiet:
+        print(f'Reading "{path}"...')
+    with path.open('r', encoding="utf-8-sig") as fh:
+        return json.load(fh)
+
+
+# Converts object to JSON format and writes it to the specified location
+def save_json_file(data: JsonData, path: Path, quiet: bool = False) -> None:
+    with path.open('w') as file:
+        if not quiet:
+            print(f'Writing "{path}"...')
+        json.dump(data, file, indent=2)
+
+
+# Load a text file into a string list with an entry for each line
+# Removes newlines from end of each line
+def load_text_file(path: Path, quiet: Optional[bool] = False) -> list[str]:
+    if not path.exists():
+        sys.exit(f'{path} not found, exiting')
+    if not quiet:
+        print(f'Reading "{path}"...')
+    return path.read_text().splitlines()
+
+
+# Overwrite file with each string in lines list adding newline to each
+def save_text_file(lines: list[str], path: Path, quiet: Optional[bool] = False) -> None:
+    if not quiet:
+        print(f'Writing "{path}"...')
+    with path.open('w') as file:
+        for line in lines:
+            file.write(line + '\n')
+
+
+# This file is generated using hard coded information and info from a save
+# file. It is generated by dumpsavefile.js
+def read_savedpadpipes(game: str, sourcedir: Path) -> dict[str, list]:
+    path = sourcedir.joinpath(f'savedpadpipes.{game}.json')
+
+    # Open and read the whole js file if it exists
+    if path.exists():
+        print(f'Reading {path}...')
+        return load_json_file(path=path)
+
+    print(f'Warning: {path} not found, skipping')
+    return {'pipes': [], 'pads': []}
+
+
+# Load the filelist we have extracted from the specific game which gives the
+# path of all files, so that we can look for files corresponding to enums and
+# structures which don't include the path in the object instance
+def load_filelist(path: Path, quiet: Optional[bool] = False):
+    lines = load_text_file(path=path.joinpath('gamefilelist.txt'), quiet=quiet)
+    basedict = {}
+    for line in lines:
+        basename, ext = line.split('/')[-1].split('.')
+
+        # We're only interested in uasset files (for bps, enums etc)
+        if ext.startswith('uasset'):
+            # Add line to array for each base filename
+            if not basedict.get(basename):
+                basedict[basename] = []
+            basedict[basename].append(line)
+
+    return {'lines': lines, 'basedict': basedict}
+
+
+# Write an asset list for use by CUE4Parse to select which files to extract
+# from the game PAK file. Blueprints needed for localisation and enums just
+# for understanding data
+def save_assetlist(  # noqa: C901 - disable complexity warning
+    items: list[str], filelist: dict, path: Path, quiet: Optional[bool] = False, prefer: Optional[str] = None
+) -> None:
+
+    def after_rfind_nth(s: str, ss: str, n: int):
+        idx = len(s)
+        for i in range(n):
+            idx = s.rfind(ss, 0, idx)
+        return s[idx if idx >= 0 else 0 :]
+
+    lines = []
+    basedict = filelist['basedict']
+    for item in items:
+        if item.find('/') == -1:
+            if matching_lines := basedict.get(item):
+                match_idx = 0
+                if len(matching_lines) > 1:
+                    if prefer:
+                        for i, line in enumerate(matching_lines):
+                            if prefer + '/' + item + '.uasset' in line:
+                                match_idx = i
+                    print(
+                        f'Warning: {len(matching_lines)} options for {item} '
+                        f'in gamefilelist.txt using #{match_idx} '
+                        f'(...{after_rfind_nth(matching_lines[match_idx], '/', 3)})'
+                    )
+                lines.append(matching_lines[match_idx])
+        else:
+            names = item.split('.')[0].split('/')
+            if matching_lines := basedict.get(names[-1]):
+                fullname = '/'.join(names[1:]) + '.uasset'
+                for line in matching_lines:
+                    if line.lower().endswith(fullname.lower()):
+                        lines.append(line)
+    lines.sort()
+    save_text_file(lines=lines, path=path, quiet=quiet)
+
+
+# Read properties from the blueprints for the specified set of property keys
+def load_blueprint_keys(sourcedir: Path, def_keys: list[str], type_keys: dict[str, str] = None):
+    if type_keys is None:
+        type_keys = {}
+    path = sourcedir.joinpath('bp')
+    bp_defaults = {}
+    for filename in path.glob('*.json'):
+        jsondata = load_json_file(path=filename, quiet=True)
+        otype = filename.stem + '_C'
+        bp_defaults[otype] = {}
+        for obj in jsondata:
+            keys = def_keys if obj['Type'] == otype else type_keys.get(obj['Type'])
+            if keys and (p := obj.get('Properties')):
+                bp_defaults[otype] |= {key: p[key] for key in keys if key in p}
+        if not bp_defaults[otype]:
+            del bp_defaults[otype]
+
+    return bp_defaults
