+++ scripts/supraland_parser.py
@@ -1,2247 +1,13 @@
-import json
-import re
-import sys
 from argparse import ArgumentParser
-from itertools import groupby
-from math import radians
 from os import environ
 from pathlib import Path
 from typing import Any, Optional
 
-import networkx as nx
-import numpy as np
-from libpysal import weights
-from mathutils import Euler, Matrix, Quaternion, Vector
-from PIL import Image
-from sklearn.neighbors import KDTree
-
-try:
-    import win32com.client
-except ModuleNotFoundError:
-    pass
-
-type JsonData = Any
-
-'''
-References to other UE objects the level files and blueprints often use a pair
-of strings:
-    AssetNamePath / SubStringPath
-or  ObjectName / ObjectPath
-
-These are identifying a file from the PAK (to be found in gamefilelist.txt)
-by a filename, possibly an object index etc.
-
-When AssetPathName and ObjectPath are used to reference a file that will be
-found in the gamefilelist.txt. The first word in the path may be a mapping
-name as follows:
-
-In Supraland / Crash
-[/]Game/        -> Supraland/Content
-
-In SIU
-[/]Game/        -> SupralandSIU/Content
-
-In SW
-[/]Game/        -> Supraworld/Content
-[/]Supraworld/  -> Supraworld/Plugins/GameFeatures/Supraworld/Supraworld/Content/
-[/]SupraAssets/ -> Supraworld/Plugins/Supra/SupraAssets/Content/
-[/]SupraCore/   -> Supraworld/Plugins/Supra/SupraCore/Content/
-
-The path may be relative to any of these directories or there may be no
-path.
-
-For objects in maps, AssetPathName will be map file reference and SubStringPath
-will be the object name hierarchy ie PersistentLevel.{ObjectName}
-
-For blueprint references AssetPathName will be the bluerint file with .{type_C}
-and SubStringPath may be empty
-
-For ObjectName/ObjectPath ObjectPath will have object path to a file and the index
-of the object within that file. ObjectName gives the name path (outer.outer.name)
-
-Or ObjectName can contain a class and ObjectPath the place where that is defined
-using the map.index method
-
-So we can normally decode the reference to:
-1. A type / class of the object being referenced
-2. A physical file where the relevant definition exists
-3. A way to find the specific object within the file
-
-
-# RootComponent":
-#    "ObjectName": "SceneComponent'Map:PersistentLevel.PipesystemNew10.DefaultSceneRoot'", type'area:*.*.outer.name'}
-#    "ObjectPath": "Supraland/Content/FirstPersonBP/Maps/Map.28893" /area.index
-# Pickup Class
-#    "ObjectName": "BlueprintGeneratedClass'Pickup_SolversGuidePage_C'",    -> 'class$
-#    "ObjectPath": "/Supraworld/Levelobjects/Collectible/Pickup_SolversGuidePage.0" not a map object just a class
-# RequiredAbilities (array)
-#    "AssetPathName": "/Supraworld/Abilities/Dash/Inventory_Dash.Inventory_Dash_C",
-#    "SubPathString": ""
-# InventoryItem
-#    "AssetPathName": "/Supraworld/Abilities/ThoughtReading/Inventory_ThoughtReading.Inventory_ThoughtReading_C",
-#    "SubPathString": ""
-# CustomShopItem
-#    "AssetPathName": "/Supraworld/Systems/Shop/ShopItems/ShopItemAbility_MindReading.ShopItemAbility_MindReading_C",
-#    "SubPathString": ""  <- It's not a map obect, so basically just extract the class ie .class$
-# otherPipeInOtherLevel: {
-#    "AssetPathName": "/Game/FirstPersonBP/Maps/DLC2_Area0_Below.DLC2_Area0_Below", -> /area.area$
-#    "SubPathString": "PersistentLevel.PipesystemNew_IntoSewer" -> Outer.Name$
-
-    ObjectName:
-        "BlueprintGeneratedClass'{type}'"                   <-- blueprint class
-        "{type}'{file}:PersistentLevel.{outer}.{name}'"     <-- instance (file can be map)
-        "{type}'Default__{type}[:name]'"                    <-- templates (name is child of default object)
-        "Function'{class}:{name}
-
-    Seems like the thing between ' and : is the object this belongs to a map, a class, a default structure
-    it basically seems to be what the thing is in then the bit after : is the hierarchy relevant to it
-    so it might be a class or it might be a level or some object
-
-    ObjectPath:
-        map/blueprint file reference (map => area) using index into array of objects
-        "[/][symboliclink/]{file path}/{file}.{index}
-
-    for this: ObjectPath always gives the file the thing is contained in and index which object (for the definition)
-    and ObjectName the last bit between one of ': and the final ' will be the name of the object. If the object is
-    the definition of a class it's a type (generally ends in _C)
-
-    Used for references to classes (to create one) or a specific instance in another file ()
-    AssetPathName:
-        map/blueprint file reference using name of object (SubPathString if not unique)
-        "[/][symboliclink/]{file path}/{blueprint or map file}.name"  <-- if subPathString is empty {name} is class otherwise map area
-        "" or "None"                                                 <-- No ref
-    SubPathString:
-        Unique hierarchy names to a specific object in a file (grandparent.parent.child}
-        "PersistentLevel[.{outer}].{name}"                  <-- instance in another file
-        ""                                                  <-- class
-
-    for this: AssetPathName gives the file along with map or class. SubPathString
-    If non-unique subpathstring clarifies with the full hierarchy. So required for instances
-    not required for class/type
-
-    Instance:
-        area       Map, Crash, DLC2_*, Supraworld
-        file       [/][symboliclink/]{file path}/{bp or map name}{.uasset|.umap}
-        ref        index of object in a uasset file
-        type       class/type of object *_C for game objects
-        outer      the parent (could have a whole hierarchy but most of what we care about is 1st or occasionally 2nd)
-        name       name of the object most of our objects are unique with area:name
-
-    Class:
-        file       [/][symboliclink/]{file path}/{bp or map name}{.uasset|.umap}
-        type/name  in the blueprint file it will be in the name field (file names normally exclude _C)
-        ref        index of class within blueprint (not required)
-
-    type, ref = p.get('ObjectName').split("'")[0:1]
-
-    return (type, ref.
-    def objectName(p):
-        return p.get('ObjectName', '').split('.')[-1][:]
-    def ObjectClass(p):
-        return p.get('ObjectName', '').split("'")[0]
-    def objectAreaName(p):
-        return [p.get('ObjectPath').split('/')[-1].split('.')[-2], objectName(p)]
-    def objectRef(p):
-        return p['ObjectPath'].split('/')[-1].split('.')
-    def assetClass(p):
-        return p['AssetPathName'].split('.')[-1]
-    def assetAreaName(p):
-        return p['AssetPathName'].split('.')[-1], p['SubPathString'].split('.')
-            "AssetPathName": "/Supraworld/Maps/Supraworld.Supraworld",
-            "SubPathString": "PersistentLevel.NPCManagerVolume_C_UAID_7085C2B20F0E486802_1139294404"
-    # Expects a dictionary containing SubPathString and AssetPathName or containing
-    # ObjectName and ObjectPath. If we have an object loaded from JSON corresponding
-    # to the reference object then it is returned, otherwise a string containing
-    # it's class or the name of the object. if type(getObject(p) is dict: can be
-    # used to check which has been returned.
-    def getObject(p):
-        if (sps := p.get('SubPathString')) is None:
-            if (op := p.get('ObjectPath')) is None:
-                return ""
-            if '/Maps/' in op:
-                area = op.split("/Maps/")[1].split('/')[0].split(".")[0]
-                index = int(op.split(".").pop())
-                if area in maps:
-                    object = maps[area][int(index)]
-                else:
-                    object = op.split("/").pop().split('.')[0]
-            else:
-                on = p.get('ObjectName')
-                object = on.split("'")[1]   # Just extract class
-        else:
-            apn = p.get('AssetPathName', '').split('.').pop()
-            # If the SubPathString is empty then just return the last bit of the asset path name (normally type)
-            if sps == '':
-                object = apn
-            elif not (object := objects.get(apn, {}).get(sps)):
-                # elif not (areaobjects := objects.get(apn)) or not (object := areaobjects.get(sps)):
-                # See if we have referenced object in our look up table, if not return string
-                object = apn + '.' + sps.split('.').pop()   # We could remove pop to include Outer
-
-        return object
-'''
-
-
-# p = {
-# "ObjectName": "ShopEgg_C'Supraworld:PersistentLevel.ShopEgg_C_UAID_FC3497C34610BB6D01_1671702339'",
-# "ObjectPath": "/Supraworld/Maps/Supraworld.60177"
-# }
-# Returns 'Supraworld.60177'
-def objectRefStr(p: dict[str, str]) -> str:
-    return p['ObjectPath'].split('/')[-1]
-
-
-# Returns ['Supraworld', 60177]
-def objectRef(p: dict[str, str]) -> list:
-    t = objectRefStr(p).split('.')[-2:]
-    return [t[0], int(t[1])]
-
-
-# p = {
-# "ObjectName": "ShopEgg_C'Supraworld:PersistentLevel.ShopEgg_C_UAID_FC3497C34610BB6D01_1671702339'",
-# "ObjectPath": "/Supraworld/Maps/Supraworld.60177"
-# }
-# Returns 'Supraworld:PersistentLevel.ShopEgg_C_UAID_FC3497C34610BB6D01_1671702339'
-def objectKey(p: dict[str, str]) -> str:
-    return p['Objectpath'].split("'")[1]
-
-
-# Returns a key for an object based on it's area, outer and name. Won't be useful for objects
-# deeper than root or next level in the hierarchy
-def makeObjectKey(area: str, outer: str, name: str) -> str:
-    key = area + ':PersistentLevel.'
-    if outer != 'PersistentLevel':
-        key += outer
-    return key + '.' + name
-
-
-# Data about each game based on the game code sl, slc, siu, sw
-#
-# appid     Steam application id
-# pathmap   gives the mapping for file references that is used
-#           by ObjectName/ObjectPath and AssetPathName/SubPathString
-# maps      gives the name of the maps that we're actually interested in
-config = {
-    'sl': {
-        'appids': ['813630'],
-        'basename': 'Supraland',
-        'pathmap': {
-            'Game': 'Supraland/Content',
-        },
-        'maps': ['Map'],
-    },
-    'slc': {
-        'appids': ['813630'],
-        'basename': 'Supraland',
-        'pathmap': {
-            'Game': 'Supraland/Content',
-        },
-        'maps': ['Crash'],
-    },
-    'siu': {
-        'appids': ['1522870'],
-        'basename': 'SupralandSIU',
-        'pathmap': {
-            'Game': 'SupralandSIU/Content',
-        },
-        # Note: we don't include DLC2_Menu_Splash anymore as
-        # it we find one secret and one pipe both of which are not usable
-        # DLC2_LateChanges was added more recently
-        'maps': [
-            'DLC2_Complete',
-            'DLC2_FinalBoss',
-            'DLC2_Area0',
-            'DLC2_SecretLavaArea',
-            'DLC2_PostRainbow',
-            'DLC2_Area0_Below',
-            'DLC2_RainbowTown',
-            'DLC2_Splash',
-            'DLC2_Menu',
-            'DLC2_LateChanges',
-        ],
-    },
-    'sw': {
-        'appids': ['1869290', '1869291'],
-        'basename': 'Supraworld',
-        'pathmap': {
-            'Game': 'Supraworld/Content',
-            'Supraworld': 'Supraworld/Plugins/GameFeatures/Supraworld/Supraworld/Content/',
-            'SupraAssets': 'Supraworld/Plugins/Supra/SupraAssets/Content/',
-            'SupraCore': 'Supraworld/Plugins/Supra/SupraCore/Content/',
-        },
-        'maps': ['Supraworld'],
-    },
-}
-
-
-# Early Access Filter data
-ea_filter = True
-ea_fogfile = "swmapfog-ea.png"
-ea_proggroups = [
-    'Act1',
-    'Act2.Blue',
-    'Act2.Green',
-]
-ea_areas = [
-    'ArmChairTown',
-    'BedDrawer',
-    'BedKingdom',
-    'Castle',
-    'OutskirtsCastle',
-    'OutskirtsStartTown',
-    'RecyclingArea',
-    'SnailArea',
-    'StartTown',
-]
-ea_abilities = [
-    'BlowGun',
-    'Crouch',
-    'Dash',
-    'Jump',
-    'JumpHigh',
-    'Vision',
-    'SpongeSuit',
-    'Toothpick',
-    'Dart',
-    'Stake',
-    'Strength',
-]
-ea_fog_bounds = (-116500, -116500, 83500, 83500)
-ea_fog_pixels = None
-ea_fog_width = 0
-ea_fog_height = 0
-
-
-# Classes that potentially have a travel target
-travel_types = [
-    'Balloon_C',
-    'DashLauncher_C',
-    'GuardVolume_C',
-    'Jumppad_C',
-    'Jumppad_TwoPath_C',
-    'LaunchBox_C',
-    'LaunchVolume_C',
-    'RubberBand_C',
-    'RubberBand_CustomDeform_C',
-    'WobbleSpring_C',
-]
-
-# Override Material Variants
-material_types = ["Nailscrew_C", "PuzzleCloud_C"]
-
-tracked_staticmeshes = ['Poker_Chip_1']
-
-staticmesh2variant = {'Poker_Chip_1': 'red'}
-
-material2variant = {
-    'Metal_Base_Golden': 'gold',
-    'CloudPurple': 'purple',
-    'CloudCyan': 'cyan',
-    'CloudRed': 'red',
-    'CloudBlue': 'blue',
-    'CloudWhite': 'white',
-    'Poker_Chip_Base_Red': 'red',
-    'Poker_Chip_Black': 'black',
-    'Poker_Chip_Blue': 'blue',
-    'Poker_Chip_Brown': 'brown',
-    'Poker_Chip_Green': 'green',
-    'Poker_Chip_Pink': 'pink',
-    'Poker_Chip_Purple': 'purple',
-    'Poker_Chip_Yellow': 'yellow',
-    'Poker_Chip_YellowBlack': 'yellowblack',
-}
-
-
-dietype2typeprefix = {
-    'DieType::D6': 'D6',
-    'DieType::D6 Round': 'D6R',
-}
-
-# Currently there are three Loot Pools:
-#
-# Lootpool1: Inventory_BlowgunBlowTime1_C
-# Lootpool2: Inventory_DashDamage_C, Inventory_DashBurstRange_C, Inventory_ToothpickDamage_C, Inventory_MaxHealth_C
-# Lootpool3Money: Inventory_Coin3_C
-
-marker_types = {
-    'PlayerStart',
-    'Jumppad_C',
-    'Bones_C',
-    'Chest_C',
-    'BarrelColor_C',
-    'BarrelRed_C',
-    'Battery_C',
-    'BP_A3_StrengthQuest_C',
-    'Lift1_C',
-    'DeadHero_C',
-    'ExplodingBattery_C',
-    'GoldBlock_C',
-    'GoldNugget_C',
-    'Jumppillow_C',
-    'MoonTake_C',
-    'Plumbus_C',
-    'Stone_C',
-    'ValveCarriable_C',
-    'ValveSlot_C',
-    'Valve_C',
-    'MatchBox_C',
-    'Shell_C',
-    'BarrelClosed_Blueprint_C',
-    'MetalBall_C',
-    'Supraball_C',
-    'Key_C',
-    'KeyLock_C',
-    'KeycardColor_C',
-    'PipeCap_C',
-    'Sponge_C',
-    'Juicer_C',
-    'Seed_C',
-    'Anvil_C',
-    'Map_C',
-    'NomNomFlies_C',
-    'CarrotPhysical_C',
-    'RingColorer_C',
-    'RespawnActor_C',
-    'CarryStones_Heavy_C',
-    'CarryStones_C',
-    'Crystal_C',
-    'RingRusty_C',
-    'SecretFound_C',
-    # slc
-    'Scrap_C',
-    'TalkingSpeaker_C',
-    'Sponge_Large_C',
-    # siu
-    'HealingStation_C',
-    'BP_EngagementCup_Base_C',
-    'SlumBurningQuest_C',
-    'Trash_C',
-    'BP_Area2_Uncloged_Quest_C',
-    'BathGuyVolume_C',
-    'BP_A3_RobBoss_C',
-    'BP_Area2_FatGuyQuest_C',
-    'BP_ParanoidQuest_C',
-    'BP_A3_BBQ_C',
-    'BP_RebuildSlum_C',
-}
-
-starts_with = {
-    'Pipesystem',
-    'Buy',
-    'BP_Buy',
-    'BP_Purchase',
-    'BP_Unlock',
-    'Purchase',
-    'Upgrade',
-    'Button',
-    'Smallbutton',
-    'Coin',
-    'Lighttrigger',
-    'LotsOfCoins',
-    'EnemySpawn',
-    'Destroyable',
-    'BP_Pickaxe',
-    'Door',
-    'Key',
-    'ProjectileShooter',
-    'MinecraftBrick',  # can be MinecraftBrick_C and MinecraftBrickRespawnable_C
-    'CrashEnemySpawner_C',
-}
-
-ends_with = {
-    'Chest_C',
-    'Button_C',
-    'Lever_C',
-    'Meat_C',
-    'Loot_C',
-    'Detector_C',
-    'Door_C',
-    'Flower_C',
-    'Coin_C',
-    'Guy_C',
-    'TriggerVolume_C',  # opens pipes in SIU
-}
-
-properties = [
-    'IsInShop',
-    'canBePickedUp',
-    'PriceType',  # BP_UnlockMap_C, etc.
-    'Coins',
-    'CoinsInGold',
-    'Cost',
-    'Value',  # Chest_C
-    'HitsToBreak',
-    'bObsidian',
-    'HitsTaken',
-    'BrickType',  # Minecraftbrick_C
-    'AllowEnemyProjectiles',
-    'RequiresPurpleShot?',
-    'ButtonType',
-    'Shape',  # Button_C
-    'Color',
-    'OriginalColor',  # Seed_C/*Flower_C/Keycard*_C (0 - white, 1 - yellow, 2 - red, 5 - green)
-    'RelativeVelocity',
-    'AllowStomp',
-    'DisableMovementInAir',
-    'RelativeVelocity?',
-    'CenterActor',  # jumpppad_c
-    'Achievement?',
-    'Achievement Name',  # trigger volumes
-    'Contains Coin',  # DestroyablePots_C
-    'bDoesntRotate',  # Coin_C, CoinBig_C
-    'Scrapamount',  # Scrap_C
-]
-
-
-# Unreal mostly uses PascalCase / UpperCamelCase for its naming. This function converts a string of this form
-# to snake case ie all lower case with underscores. It has special handling for flags converting 'bMyVar' to
-# 'is_my_var' and 'MyVar?' to 'my_var_flag'
-def camel_to_snake(s: str) -> str:
-    if s[-1] == '?':
-        s = s[:-1] + 'Flag'
-    if s.startswith('b'):
-        s = 'Is' + s[1:]
-    s = s.replace(' ', '')
-    return ''.join(['_' + c.lower() if c.isupper() else c for c in s]).lstrip('_')
-
-
-# Returns list of numbers in the specified string (ie "a1b98" return ["1", "98"])
-def get_ints(s: str) -> list[str]:
-    return [''.join(group) for key, group in groupby(s, lambda e: e.isdigit()) if key]
-
-
-# Returns last number in the specified string
-def get_last_int(s: str) -> str:
-    return ints[-1] if (ints := get_ints(s)) else ''
-
-
-# Returns number at end of the specified string or empty string if it isn't one
-# Note: we could do this more simply with a regexp but it seems inefficient
-# def get_last_int(s):
-#     return m.group() if (m := re.search('\d+$', s)) else ''
-def get_end_int(s: str) -> str:
-    return get_ints(s)[-1] if s and s[-1].isdigit() else ''
-
-
-# Return True if this string looks like some kind of enumeration
-def isenum(s: Any) -> bool:
-    return (
-        isinstance(s, str)
-        and '::' in s
-        and set(s) <= set("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz:0123456789")
-    )
-
-
-# Return True if this string looks like an Unreal renamed enumeration
-def isueenum(s: Any) -> bool:
-    return isinstance(s, str) and '::NewEnumerator' in s
-
-
-class UEEnums:
-    def __init__(self):
-        self.map = {}
-        self.types = set()
-
-    def loadenumbp(self, path: Path):
-        # Read the file as a JSON
-        j = load_json_file(path=path, quiet=True)
-
-        # Check that it is formatted as we expect
-        if (
-            not isinstance(j, list)
-            or not j  # We have a non empty list
-            or not isinstance(j[0], dict)  # containing a dictionary
-            or not j[0].get('Type') == 'UserDefinedEnum'
-        ):  # of the right type
-            print(f"Warning: {path[-1]} does not appear to be an Unreal enumeration")
-            return
-
-        name = j[0]['Name']
-
-        for e in j[0]['Properties']['DisplayNameMap']:
-            k, v = e['Key'], e['Value']
-            if (s := v.get('SourceString')) or (s := v.get('CultureInvariantString')):
-                if s != k:
-                    self.map[name + '::' + k] = name + '::' + s
-            else:
-                print(f"Warning: in {path[-1]} {k} Value does not have SourceString nor CultureInvariantString'")
-
-        self.types.add(name)
-
-    # If s is an enumeration name we're aware of convert it to the source form
-    def ue2source(self, s: str, default: Optional[str] = None) -> str:
-        return self.map.get(s, default)
-
-    # Returns true if the two enum strings match. First may be a UE renamed enum
-    # or a source enum, second should be an untranslated source name
-    def match(self, ue: str, s: str) -> bool:
-        return ue == s or self.ue2source(ue) == s
-
-    # If first argument is a UE enum then add it
-    def addueattr(self, ue: str, s: Optional[str] = None) -> None:
-        if isueenum(ue):
-            if s and s != ue:
-                self.map[ue] = s
-            self.types.add(ue[0 : ue.find('::')])
-
-
-# Load all enumerations
-def load_all_enumbp(path: Path) -> UEEnums:
-    ueenums = UEEnums()
-    for filename in path.glob('*.json'):
-        ueenums.loadenumbp(filename)
-    return ueenums
-
-
-# Retrieve Unreal version information from the EXE file properties using
-# windows shell API. These are the properties on the file properties -> details
-# tab in explorer.
-# The API NameSpace.GetDetailsOf usually has a fixed index to properties relationship,
-# but only some seem to be officially documented, we're after the 'File version' and
-# may at some point need 'Product version'. We've now seen two different indices for
-# 'File version' so we search rather than use a hard coded index and log the info for
-# awareness.
-def get_unreal_version(exepath: Path) -> dict[str, int]:
-    uever = {}
-    sh = win32com.client.gencache.EnsureDispatch('Shell.Application', 0)
-    ns = sh.NameSpace(str(exepath.parent))
-    item = ns.Items()[exepath.name]
-
-    print(f"Retrieving UE version from '{exepath}'")
-    file_version_indices = [166, 167]
-    product_version_indices = [301]
-
-    # Search all commonly used indices. We could probably narrow this range as many are fixed
-    for i in range(350):
-        # Calling GetDetailsOf with the file name (or None) gets you the property name for index
-        prop = ns.GetDetailsOf(item.Name, i)
-        if 'version' in prop.lower():
-            # Calling GetDetailsOf with the item gets you the value of the property
-            value = ns.GetDetailsOf(item, i)
-
-            if 'product' in prop.lower():
-                print(
-                    f"'{prop}' index={i} ({value}) {'** unexpected index' if i not in product_version_indices else ''}"
-                )
-
-            if 'file' in prop.lower():
-                print(f"'{prop}' index={i} ({value}) {'** unexpected index' if i not in file_version_indices else ''}")
-                file_version = value
-
-    if not file_version:
-        print(f'Error: Failed to find file version for {exepath}')
-        exit(-1)
-
-    uever['ver'] = ns.GetDetailsOf(item, 167)
-    vernums = uever['ver'].split('.')
-    uever['major'] = int(vernums[0])
-    uever['minor'] = int(vernums[1])
-    return uever
-
-
-# Retrieves the steam branch name
-def get_steam_branch(game: str, path: Path) -> str:
-
-    # Get path to app manifest from game install path and appid
-    basepath = str(path)[0 : str(path).find('common')]
-    for appid in config[game]['appids']:
-        manifest = Path(f'{basepath}appmanifest_{appid}.acf')
-
-        # If the file exists, open it and read contents
-        if manifest.is_file():
-            print(f"Retrieving branch from '{manifest}'")
-            # Search for the current user config beta key
-            if match := re.search(r'UserConfig"[\s\S]*?"BetaKey"\s*"([^"]*)', manifest.read_text()):
-                return match.group(1)
-
-    return 'public'
-
-
-def get_swbuild_info(installdir: Path) -> dict[str, str]:
-    print(f"Retrieving build info from '{installdir}\\build.vc'")
-    with installdir.joinpath('build.vc').open('r') as file:
-        build = file.readline().rstrip()
-        date = file.readline().rstrip()
-    return {"build": build, "date": date}
-
-
-# Update Supraworld version information based on steam version installed in 'installdir'
-def update_swversion_info(game: str, datadir: Path, sourcedir: Path, mapimage: Optional[str] = None) -> None:
-
-    versions = load_json_file(path=datadir.joinpath('versions.json'))
-
-    # versions JSON looks something like this:
-    # {
-    #     ...
-    #     "sw": {
-    #         "ue": {"major": 5, "minor": 6, "ver": "5.6.1.0" },
-    #         "game": {"version": "EarlyAccess", "build": 10596, "date": "08/04/2026 - 13:37 UTC", "branch": "public", "type": "Shipping", "mapver": "V8" }
-    #     }
-    # }
-
-    basename = config['sw']['basename']
-    exe_file = Path(list(sourcedir.joinpath(basename, 'Binaries\\Win64').glob(f"{basename}-Win64*.exe"))[0])
-
-    uever = get_unreal_version(exepath=exe_file)
-    versions['sw']['ue'].update(uever)
-
-    gamever = get_swbuild_info(installdir=sourcedir)
-    gamever['branch'] = get_steam_branch(game='sw', path=sourcedir)
-    gamever['type'] = exe_file.stem.split('-')[-1]
-
-    if mapimage:
-        gamever['mapver'] = 'V' + str(get_ints(mapimage)[0])
-
-    versions['sw']['game'].update(gamever)
-
-    save_json_file(data=versions, path=datadir.joinpath('versions.json'))
-
-    print("SW Version: ", f'{gamever | {'ue': uever['ver']}}')  # version, build, date, branch, type, mapver
-
-
-# We presume export.cmd has been used to export a set of map files for the game to
-# a sub-directory of 'sourcedir'. We are going to go through the map and generate
-# information about the
-def preproc_levels(game: str, datadir: Path, sourcedir: Path) -> None:  # noqa: C901 - disable complexity warning
-
-    # Load in any enumerations we have found / extracted
-    ueenums = load_all_enumbp(path=sourcedir.joinpath('enums'))
-
-    # Load current gameClasses.json so we know which classes we currently know about for this game
-    game_classes = load_json_file(path=datadir.joinpath('gameClasses.json'))
-
-    # Load the game's PAK file list so we can look up where to find enums
-    gamefilelist = load_filelist(path=sourcedir)
-
-    # Set of blueprint CUE4Parse asset paths for types in gameClasses.json
-    # and set of other classes which end in _C
-    bp_assetlist = set()
-    area_names = set()
-    ignored_types = set()
-    classprops = {}
-
-    # Set of property:type pairs for all key:value pairs in game maps
-    propset = {'Root': set(), 'Properties': set(), 'Other': set()}
-
-    # Loop through all the level json's we've extracted
-    for filename in sourcedir.joinpath('levels').glob('*.json'):
-        # Area name is everything between last '\' and the '.json'
-        filestr = str(filename)
-        area = filestr[filestr.rfind('\\') + 1 : -5]
-        area_names.add(area)
-
-        # Read the level file in and loop through the outer list of objects)
-        level = load_json_file(path=filename)
-        level_changed = False
-        for obj in level:
-            if (
-                not (otype := obj.get('Type'))
-                or not (oname := obj.get('Name'))
-                or not (obp := obj.get('Class').split("'")[1].split('.')[0])
-            ):
-                print(f'Warning: Object missing something in {filename} Type:{otype} Name:{oname} BP:{obp}')
-                continue
-
-            # Collect any classes which use the properties we're interested
-            if p := obj.get('Properties'):
-                for prop in [
-                    'RequiredAbilities',
-                    'Area',
-                    'ProgressionGroup',
-                    'AdditionalRequirementHints',
-                    'AdditionalRequirements',
-                    'DieType',
-                    'Value',
-                    'CoinValue',
-                    'Coins',
-                    'CoinPool',
-                    'Cost',
-                    'Pickup Class',
-                    'CustomShopItem',
-                    'InventoryItem',
-                    'Initial Shop Inventory',
-                    'SupraworldLaunchComponent',
-                    'FrontSupraworldLaunchComponent',
-                    'BackSupraworldLaunchComponent',
-                    'AltLaunchComp',
-                    'SpawnerTags',
-                    'Spawn on Level Start',
-                    'DisplayName',
-                    'DisplayDescription',
-                ]:
-                    if prop in p:
-                        classprops[otype] = set([*list(classprops.get(otype, set())), prop])
-
-            # Remember blueprint paths of types we're interested in
-            # And the base type of any other custom types
-            # Note: old SL based game classes don't necessarily have games member
-            if otype in game_classes and game in game_classes[otype].get('games', ['sl', 'slc', 'siu']):
-                bp_assetlist.add(obp if obp[0] != '/' else obp[1:])
-                p = obj['Properties']
-                # Collect those classes we're interested that use these properties
-                for prop in [
-                    'Color',
-                    'Color_Initial',
-                    'Initial_Color',
-                    'ButtonColor',
-                    'LiquidColor',
-                    'RuneColor',
-                    'bHidden',
-                    'bHiddenInGame',
-                    'bInitialExists',
-                    'bExists',
-                    'Spawn on Level Start',
-                    'bItemIsAvailable_Initial',
-                ]:
-                    if prop in p:
-                        classprops[otype] = set([*list(classprops.get(otype, set())), prop])
-            elif otype.endswith('_C'):
-                ignored_types.add(otype)
-
-            # Walk all data in the level remembering property names/types
-            # Any enum types used and if possible remap enum attributes to source names
-            def gather_properties(obj: dict, setkey: str):
-                nonlocal level_changed
-
-                for ref, value in obj.items() if isinstance(obj, dict) else enumerate(obj):
-                    proptype = type(value).__name__
-                    if isenum(value):
-                        proptype = value[0 : value.find('::')]
-                        ueenums.addueattr(value)
-                        if isueenum(value):
-                            if source := ueenums.ue2source(value):
-                                obj[ref] = source
-                                level_changed = True
-
-                    if isinstance(obj, dict):
-                        otype = None
-                        if ref == 'ObjectName' and value.startswith('BlueprintGeneratedClass'):
-                            otype = value.split("'")[1]
-                            obp = obj['ObjectPath'].split('.')[0]
-                        if ref == 'AssetPathName' and value and '.' in value:
-                            obp, otype = value.split('.')
-                        if otype:
-                            if (
-                                otype == 'Jumppad_C'
-                                or otype in game_classes
-                                and game in game_classes[otype].get('games', ['sl', 'slc', 'siu'])
-                            ):
-                                bp_assetlist.add(obp if obp[0] != '/' else obp[1:])
-                            elif otype.endswith('_C'):
-                                ignored_types.add(otype)
-                            continue
-                        propset[setkey].add(ref + ':' + proptype)
-
-                    if isinstance(value, (dict, list)):
-                        gather_properties(obj=value, setkey='Properties' if ref == 'Properties' else 'Other')
-
-            gather_properties(obj=obj, setkey='Root')
-
-        # If we changed this level map's enumerations then write it out again
-        if level_changed:
-            save_json_file(data=level, path=filename)
-
-    save_assetlist(items=bp_assetlist, filelist=gamefilelist, path=sourcedir.joinpath('bpassetlist.txt'))
-    save_assetlist(
-        items=ueenums.types, filelist=gamefilelist, path=sourcedir.joinpath('enumassetlist.txt'), prefer="Enums"
-    )
-    save_text_file(lines=sorted(area_names), path=sourcedir.joinpath("areanames.txt"))
-
-    for k, v in classprops.items():
-        classprops[k] = sorted(list(v))
-    levelprops = {
-        "ClassProps": dict(sorted(classprops.items())),
-        "IgnoredTypes": sorted(list(ignored_types)),
-        "EnumTypes": sorted(list(ueenums.types)),
-        "EnumValues": ueenums.map,
-        "RootProps": sorted(list(propset['Root'])),
-        "Properties": sorted(list(propset['Properties'])),
-        "OtherProps": sorted(list(propset['Other'])),
-    }
-    save_json_file(data=levelprops, path=sourcedir.joinpath('levelprops.json'))
-
-
-# Load the filelist we have extracted from the specific game which gives the
-# path of all files, so that we can look for files corresponding to enums and
-# structures which don't include the path in the object instance
-def load_filelist(path: Path, quiet: Optional[bool] = False):
-    lines = load_text_file(path=path.joinpath('gamefilelist.txt'), quiet=quiet)
-    basedict = {}
-    for line in lines:
-        basename, ext = line.split('/')[-1].split('.')
-
-        # We're only interested in uasset files (for bps, enums etc)
-        if ext.startswith('uasset'):
-            # Add line to array for each base filename
-            if not basedict.get(basename):
-                basedict[basename] = []
-            basedict[basename].append(line)
-
-    return {'lines': lines, 'basedict': basedict}
-
-
-# Write an asset list for use by CUE4Parse to select which files to extract
-# from the game PAK file. Blueprints needed for localisation and enums just
-# for understanding data
-def save_assetlist(  # noqa: C901 - disable complexity warning
-    items: list[str], filelist: dict, path: Path, quiet: Optional[bool] = False, prefer: Optional[str] = None
-) -> None:
-
-    def after_rfind_nth(s: str, ss: str, n: int):
-        idx = len(s)
-        for i in range(n):
-            idx = s.rfind(ss, 0, idx)
-        return s[idx if idx >= 0 else 0 :]
-
-    lines = []
-    basedict = filelist['basedict']
-    for item in items:
-        if item.find('/') == -1:
-            if matching_lines := basedict.get(item):
-                match_idx = 0
-                if len(matching_lines) > 1:
-                    if prefer:
-                        for i, line in enumerate(matching_lines):
-                            if prefer + '/' + item + '.uasset' in line:
-                                match_idx = i
-                    print(
-                        f'Warning: {len(matching_lines)} options for {item} '
-                        f'in gamefilelist.txt using #{match_idx} '
-                        f'(...{after_rfind_nth(matching_lines[match_idx], '/', 3)})'
-                    )
-                lines.append(matching_lines[match_idx])
-        else:
-            names = item.split('.')[0].split('/')
-            if matching_lines := basedict.get(names[-1]):
-                fullname = '/'.join(names[1:]) + '.uasset'
-                for line in matching_lines:
-                    if line.lower().endswith(fullname.lower()):
-                        lines.append(line)
-    lines.sort()
-    save_text_file(lines=lines, path=path, quiet=quiet)
-
-
-# The following load/save functions make a path by joining all the varargs
-# together with '\\' and adding '.txt'
-#
-# For game specific source data this is usually something like:
-#   sourcedir:  ..\\source             (but can be any path)
-#   game:       'sl', 'siu', 'sw'      (note Crash files are in sl source directories)
-#   dir:        'levels', 'enums', 'bps', 'loc', 'mapimg'
-#   filename:   'Map'
-#
-# For output data used by the web map it's likely:
-#  '..\\public\\data', 'gameClasses'
-#
-# Extension should not be included. Empty strings will be ignored/skipped
-
-
-# Reads a json file and converts it do an object data structure
-def load_json_file(path: Path, quiet: bool = False):
-    if not path.exists():
-        sys.exit(f'{path} not found, exiting')
-    if not quiet:
-        print(f'Reading "{path}"...')
-    with path.open('r', encoding="utf-8-sig") as fh:
-        return json.load(fh)
-
-
-# Converts object to JSON format and writes it to the specified location
-def save_json_file(data: JsonData, path: Path, quiet: bool = False) -> None:
-    with path.open('w') as file:
-        if not quiet:
-            print(f'Writing "{path}"...')
-        json.dump(data, file, indent=2)
-
-
-# Load a text file into a string list with an entry for each line
-# Removes newlines from end of each line
-def load_text_file(path: Path, quiet: Optional[bool] = False) -> list[str]:
-    if not path.exists():
-        sys.exit(f'{path} not found, exiting')
-    if not quiet:
-        print(f'Reading "{path}"...')
-    return path.read_text().splitlines()
-
-
-# Overwrite file with each string in lines list adding newline to each
-def save_text_file(lines: list[str], path: Path, quiet: Optional[bool] = False) -> None:
-    if not quiet:
-        print(f'Writing "{path}"...')
-    with path.open('w') as file:
-        for line in lines:
-            file.write(line + '\n')
-
-
-sw_blueprint_keys_used = [
-    'RequiredAbilities',
-    'Area',
-    'ProgressionGroup',
-    'AdditionalRequirementHints',
-    'AdditionalRequirements',
-    'DieType',
-    'Value',
-    'CoinValue',
-    'Coins',
-    'CoinPool',
-    'Cost',
-    'Pickup Class',
-    'CustomShopItem',
-    'InventoryItem',
-    'Initial Shop Inventory',
-    'SupraworldLaunchComponent',
-    'FrontSupraworldLaunchComponent',
-    'BackSupraworldLaunchComponent',
-    'AltLaunchComp',
-    'SpawnerTags',
-    'Spawn on Level Start',
-    'DisplayName',
-    'DisplayDescription',
-    'Color',
-    'Color_Initial',
-    'Initial_Color',
-    'ButtonColor',
-    'LiquidColor',
-    'RuneColor',
-    'bHidden',
-    'bHiddenInGame',
-    'bInitialExists',
-    'bExists',
-    'Spawn on Level Start',
-    'bItemIsAvailable_Initial',
-]
-
-
-# Read properties from the blueprints for the specified set of property keys
-def load_blueprint_keys(sourcedir: Path, def_keys: list[str], type_keys: dict[str, str] = None):
-    if type_keys is None:
-        type_keys = {}
-    path = sourcedir.joinpath('bp')
-    bp_defaults = {}
-    for filename in path.glob('*.json'):
-        json = load_json_file(path=filename, quiet=True)
-        otype = filename.stem + '_C'
-        bp_defaults[otype] = {}
-        for obj in json:
-            keys = def_keys if obj['Type'] == otype else type_keys.get(obj['Type'])
-            if keys and (p := obj.get('Properties')):
-                bp_defaults[otype] |= {key: p[key] for key in keys if key in p}
-        if not bp_defaults[otype]:
-            del bp_defaults[otype]
-
-    return bp_defaults
-
-
-# Load in a PNG file containing RGB values
-def load_ea_fog(path: Path) -> bool:
-    path = path.joinpath('mapimg', ea_fogfile)
-    if not path.exists():
-        print(f'Warning: Failed to load fog png ({path})')
-        return False
-    else:
-        with Image.open(str(path)) as im:
-            global ea_fog_width, ea_fog_height, ea_fog_pixels
-            ea_fog_width = im.width
-            ea_fog_height = im.height
-            ea_fog_pixels = im.getchannel(0).load()
-        return True
-
-
-def in_earlyaccess(otype, p, pos):  # noqa: C901 - disable complexity warning
-    if not ea_filter:
-        return True
-
-    # If it has a progression and it's not in released content then reject
-    if (
-        (proggroup := p.get('ProgressionGroup', {}).get('TagName'))
-        and proggroup != 'None'
-        and not any(proggroup.endswith(s) for s in ea_proggroups)
-    ):
-        return False
-
-    # If it has an area that's not
-    if (
-        (swarea := p.get('Area', {}).get('TagName'))
-        and swarea != 'None'
-        and not any(swarea.endswith(s) for s in ea_areas)
-    ):
-        return False
-
-    # Some classes should be rejected if they have no area tag
-    if not swarea and otype in ['SecretVolume_C']:
-        return False
-
-    # Check if it requires abilities not yet released
-    if required := p.get('RequiredAbilities'):
-        for r in required:
-            if isinstance(r, str) and not any(r.endswith(s) for s in ea_abilities):
-                return False
-
-    if ea_fog_pixels:
-        xmin, ymin, xmax, ymax = ea_fog_bounds
-        if xmin < pos.x < xmax and ymin < pos.y < ymax:
-            px = (pos.x - xmin) / (xmax - xmin) * ea_fog_width
-            py = (pos.y - ymin) / (ymax - ymin) * ea_fog_width
-            if ea_fog_pixels[px, py] < 255:
-                return False
-
-    # Should probably check pos against a cuboid but not yet
-    return True
-
-
-def optEnum(s: str | int):
-    if match := re.search('::.*([0-9]+)$', str(s)):
-        return int(match.group(1))
-    return s
-
-
-def optArea(a, k, v):
-    return v if a == k else ':'.join((k, v))
-
-
-def optColor(p):
-    return p and '#' + ''.join(hex(int(p[c]))[2:] for c in 'RGB')
-
-
-def optKey(d, k, v):
-    return v is not None and d.__setitem__(k, optEnum(v))
-
-
-def getVec(d, v=0):
-    return Vector((d['X'], d['Y'], d['Z'])) if d else Vector((v, v, v))
-
-
-def getRot(d, v=0):
-    return Euler((radians(-d['Roll']), -radians(d['Pitch']), radians(d['Yaw']))) if d else Euler((v, v, v))
-
-
-def getQuat(d, v=0):
-    return Quaternion((d['W'], d['X'], d['Y'], d['Z'])) if d else Quaternion((v, v, v, v))
-
-
-def getXYZ(v):
-    return dict(x=v.x, y=v.y, z=v.z)
-
-
-def export_sw_markers(game: str, datadir: Path, sourcedir: Path):  # noqa: C901 - disable complexity warning
-    maps = {}  # dictionary from map name to json data list
-    toyeggs = {}  # Collected chocolate eggs
-    staticmeshes = {}  # dictionary from outer name to static mesh name
-    meshmats = {}  # dictionary from outer name to mesh material variant
-    targets = {}  # dictionary from outer name to target positions
-    area_mtx = {}  # Transform for each area map geometry
-    data = []  # Output marker data
-
-    # Load in any enumerations we have found / extracted
-    ueenums = load_all_enumbp(path=sourcedir.joinpath('enums'))
-
-    bp_defaults = load_blueprint_keys(
-        sourcedir=sourcedir, def_keys=sw_blueprint_keys_used, type_keys={"InvFrag_SupraworldShopItem": ["Cost"]}
-    )
-
-    # Phase 1: Read all the map json files in and build a look up table for references
-    # Also get any area/map file matrices (for streaming levels)
-    for area in config[game]['maps']:
-        # Store the map data
-        maps[area] = load_json_file(path=sourcedir.joinpath('levels', f"{area}.json"))
-
-        # Go through all objects in the map data and store lookups for later
-        for oidx, o in enumerate(maps[area]):
-            if not (outer := o.get('Outer')) or not (p := o.get('Properties')):
-                continue
-            oname = o['Name']
-            otype = o['Type']
-
-            # Keep list of static meshes by parent/outer
-            if otype == 'StaticMeshComponent':
-                if (sm := p.get('StaticMesh', {}).get('ObjectName')) and (
-                    sm := sm.split("'")[-2]
-                ) in tracked_staticmeshes:
-                    staticmeshes[outer] = sm
-                    if sm in staticmesh2variant:
-                        meshmats[outer] = staticmesh2variant[sm]
-
-                if oms := p.get('OverrideMaterials'):
-                    for om in oms:
-                        if (
-                            om
-                            and (v := om.get('ObjectName'))
-                            and (v := material2variant.get(v.split("'")[-2].split('.')[-1]))
-                        ):
-                            meshmats[outer] = v
-
-            # Keep list of Chocolate Egg interior objects
-            if otype == 'ChocolateEgg_C' and (v := p.get('ToyEgg')):
-                toyeggs[objectRefStr(v)] = o
-
-            # Keep a list of launch target positions
-            # I was filtering based on name but it's not reliable indicator of class
-            # some rubber bands have name 'StaticMashActor_...' for example
-            # and outer[0:outer.find("_C")+2] in travel_types): name
-            if (
-                otype == 'SupraworldLaunchComponent_C'
-                and (tl := p.get('TargetLocation'))
-                and not oname == 'AltLaunchComp'
-            ):  # For now remove Jumppad_TwoPath_C 2nd path
-                targets[outer] = [*targets.get(outer, []), {'x': tl['X'], 'y': tl['Y'], 'z': tl['Z']}]
-
-            # For maps that are divided into multiple files, there may be a LevelTransform for entities in that
-            # file relative to the persistent world that is handled by the streaming system. To correct for this
-            # we construct a matrix from the Translation/Rotation members if they exit
-            if (a := p.get('WorldAsset', {}).get('AssetPathName')) and (t := p.get('LevelTransform')):
-                area_mtx[a.split('.').pop()] = (
-                    Matrix.Translation(getVec(t.get('Translation'))) @ getQuat(t.get('Rotation')).to_matrix().to_4x4()
-                )
-
-    game_classes = load_json_file(path=datadir.joinpath('gameClasses.json'))
-
-    load_ea_fog(path=sourcedir)
-
-    # Phase 2: Go through all the objects which have types we're interested in
-    for area in maps:
-        for oidx, o in enumerate(maps[area]):
-            otype = o['Type']
-            oname = o['Name']
-            if not (outer := o.get('Outer')) or not (p := o.get('Properties')):
-                continue
-
-            if otype in bp_defaults:
-                p = bp_defaults[otype] | p
-
-            def getObject(p):
-                ref = objectRef(p)
-                return maps[ref[0]][ref[1]]
-
-            def get_matrix(o, matrix=Matrix.Identity(4)):
-                p = o.get('Properties', {})
-
-                if p.get('RelativeLocation'):
-                    matrix = (
-                        Matrix.LocRotScale(
-                            getVec(p.get('RelativeLocation')),
-                            getRot(p.get('RelativeRotation')),
-                            getVec(p.get('RelativeScale3D'), 1),
-                        )
-                        @ matrix
-                    )
-
-                for prop in ['RootObject', 'RootComponent', 'DefaultSceneRoot', 'AttachParent']:
-                    if p.get(prop):
-                        return get_matrix(getObject(p[prop]), matrix)
-                '''
-                    node = p.get(parent)
-                    if type(node) is dict:
-                        obj = getObject(node)
-                        if type(obj) is dict:
-                            return get_matrix(obj, matrix);
-                '''
-                return matrix
-
-            matrix = get_matrix(o)
-            if area in area_mtx:
-                matrix = area_mtx[area] @ matrix
-            pos = matrix.to_translation()
-
-            # Check for early access based on type, properties and map position
-            if not in_earlyaccess(otype, p, pos):
-                continue
-
-            # If this is a shop egg inside a chocolate egg then skip it (ChocolateEgg_C)
-            if toyeggs.get(area + '.' + str(oidx)):
-                continue
-
-            # Convert poker chip static meshes to custom Poker Chip class
-            if (v := staticmeshes.get(oname)) and v == 'Poker_Chip_1':
-                otype = '_PokerChip_C'
-
-            # Add the standard data to the object
-            data.append({'name': oname, 'type': otype, 'area': area, 'lat': pos.y, 'lng': pos.x, 'alt': pos.z})
-
-            # Hidden Flag
-            if (
-                p.get('bHidden') is True
-                or p.get('bHiddenInGame') is True
-                or p.get('bExists') is False
-                or p.get('InitialExists') is False
-                or p.get('Spawn on Level Start') is False
-                or p.get('bItemIsAvailable_Initial') is False
-            ):
-                data[-1]['hidden'] = 'true'
-
-            # Handle the Obvious Area secret
-            comment = None
-            if o.get('ActorLabel') == "ObviousAreaOuttaTown":
-                # p['Area'] = {'TagName': 'OutskirtsStartTown'}
-                comment = 'Obvious Area'
-
-            variant = ''
-            # Variants from colours or override materials
-            for colkey in ['RuneColor', 'Color', 'Color_Initial', 'LiquidColor', 'ButtonColor']:
-                if (color := p.get(colkey)) and isinstance(color, str):
-                    color = ueenums.ue2source(color, color)
-                    variant = color.removeprefix("ESupraColors::").lower()
-                    break
-
-            # Currently gold screws, puzzle cloud's and poker chips
-            if v := meshmats.get(oname):
-                variant = v
-
-            if otype == 'SupraworldPlayerStart_C':
-                variant = 'red'
-
-            # Only keep gold variant of Nailscrew_C
-            if otype == 'Nailscrew_C' and variant != 'gold':
-                del data[-1]
-                continue
-
-            if variant:
-                data[-1]['variant'] = variant
-
-            # If it's a D6 or D6 Round then change the class (could use variant but might have coloured die in future)
-            if otype == 'Die_C' and (v := dietype2typeprefix.get(p.get('DieType'))):
-                data[-1]['type'] = v + ':' + otype
-
-            if otype == 'HayGuy_C' and p.get('Collectible Tag', {}).get('TagName') == 'Stats.Collectible.Thread':
-                data[-1]['type'] = '_ThreadGuy_C'
-
-            # Solve clashes with old game classes
-            if otype in ['Jumppad_C', 'KeyPlastic_C']:
-                data[-1]['type'] = 'SW:' + otype
-
-            # Grab the Area (or AreaTag) if there is one
-            if v := p.get('Area', p.get('AreaTag', {})).get('TagName'):
-                data[-1]['area_tag'] = v.split('.')[-1]
-
-            # Grab progression tag
-            if (v := p.get('ProgressionGroup', {}).get('TagName')) and v != "None":
-                data[-1]['prog_tag'] = v.removeprefix('Supraworld.Story.').replace('.', ':')
-
-            # Grab secret required abilities if there are any
-            if otype == 'SecretVolume_C' and (v := p.get('RequiredAbilities')):
-                data[-1]['abilities'] = ','.join(
-                    [a.removeprefix('GameplayAbilitySystem.Ability.').replace('.', ' ') for a in v]
-                )
-
-            # If this is a chocolate egg then get the spawn property from the related toy egg (which will not be included)
-            spawn_props = p
-            if otype == 'ChocolateEgg_C' and (v := p.get('ToyEgg')):
-                spawn_props = getObject(v)['Properties']
-
-            # Spawners (presents, shops, eggs, launch boxes, ...)
-            spawns = ''
-            if v := spawn_props.get('Pickup Class'):
-                spawns = v['ObjectName'].split("'")[-2]
-            elif (
-                (v := spawn_props.get('InventoryItem'))
-                or (
-                    (v := spawn_props.get('Initial Shop Inventory')) and isinstance((v := list(v[0].values())[0]), dict)
-                )
-                or (v := spawn_props.get('CustomShopItem'))
-            ):
-                spawns = v['AssetPathName'].split(".")[-1]
-            if spawn_props.get('bFromLootPool'):
-                spawns = '_LootPool_C'
-
-            # Cost
-            if (v := bp_defaults.get(spawns, {}).get('Cost')) and otype in [
-                'ShopItemSpawner_C',
-                'ShopEgg_C',
-                'ChocolateEgg_C',
-                'ShopSlot_C',
-            ]:
-                data[-1]['cost'] = v
-
-            # These classes just spawn stuff so we change the type to what it spawns
-            if otype in [
-                'ItemSpawner_C',
-                'PickupSpawner_C',
-                'ShopItemSpawner_C',
-                'RespawnablePickupSpawner_C',
-                'ShopSlot_C',
-            ]:
-                data[-1]['type'] = otype = spawns
-                spawns = ''
-
-            # Coins
-            # Anything that spawns Inventory_Coin[nn]_C or RealCoinPickup_C/5Cent_C/Gumball_Machine_C
-            coins = 0
-            if (v := coin_defaults.get(otype)) or (v := coin_defaults.get(spawns)):
-                coins = v
-            if (v := p.get('Value')) and v.startswith('CoinValue::'):  # RealCoinPickup_C/5Cent_C
-                coins = int(get_end_int(ueenums.ue2source(v, v)))
-            if v := p.get('CoinPool'):  # Gumball_Machine_C
-                coins = v
-            if coins:
-                data[-1]['coins'] = coins
-                if spawns != '':
-                    data[-1]['type'] = 'Coin:' + data[-1]['type']
-                    spawns = ''
-
-            # Loot boxes default to spawning loot pools
-            if otype == 'PresentBox_Lootpools_C' and not (coins or spawns):
-                spawns = '_LootPool_C'
-
-            if spawns:
-                data[-1]['spawns'] = spawns
-
-            elif otype in ['ShopEgg_C', 'ChocolateEgg_C']:
-                # Remove any empty eggs
-                del data[-1]
-                continue
-
-            def filter_targets(objpos, targets, mindist):
-                ov = Vector((objpos['lng'], objpos['lat'], objpos['alt']))
-                keep = [
-                    target
-                    for target in targets
-                    if (Vector((target['x'], target['y'], target['z'])) - ov).magnitude >= mindist
-                ]
-                return keep
-
-            # Travel targets
-            if oname in targets and (v := filter_targets(data[-1], targets[oname], 600)):
-                data[-1]['targets'] = v
-                data[-1]['linetype'] = 'target'
-
-            # Type may have been modified, so only check for it at the end
-            otype = data[-1]['type']
-
-            def class_supported(otype, game):
-                return (gc := game_classes.get(otype)) and gc.get('layer') and game in gc.get('games', [])
-
-            if not class_supported(otype, game) or spawns and not class_supported(spawns, game):
-                del data[-1]
-                continue
-
-            if comment:
-                data[-1]['comment'] = comment
-
-    save_json_file(data=data, path=datadir.joinpath(f'markers.{game}.json'))
-    print("Done")
-
-
-# Read's gameClasses.json and then tries to find friendly names and descriptions and their
-# corresponding loc keys by reading the corresponding blueprint files looking for the strings
-# It avoids overwriting values in gameClasses
-#
-# SIU and SL share class names, so we use the same strings for both, which one is chosen depends
-# on which is read first (traditionally SL then SIU)
-#
-# For some things we didn't find the strings/keys in the blueprint but hand added the values
-# by finding the strings in the loc data.
-#
-# Further edits were made to the strings/keys to make a few fixups to the english
-def export_class_loc(game: str, datadir: Path, sourcedir: Path) -> None:  # noqa: C901 - disable complexity warning
-
-    game_classes = load_json_file(path=datadir.joinpath('gameClasses.json'))
-
-    classes = set()
-    propcount = 0
-    propmatches = 0
-
-    def optKeyLoc(class_name: str, k: str, v):
-        nonlocal propcount, propmatches
-        if v:
-            if class_name not in game_classes:
-                game_classes[class_name] = {}
-            classes.add(class_name)
-            propcount += 1
-            if k not in game_classes[class_name]:
-                print(f'{class_name}.{k} => "{v}"')
-                game_classes[class_name][k] = v
-            elif game_classes[class_name][k] == v:
-                propmatches += 1
-            else:
-                print(f'{class_name}.{k} == "{game_classes[class_name][k]}" != "{v}"')
-
-    def is_class_used(cls_name):
-        return cls_name in game_classes and game in game_classes[cls_name].get('games', ['sl', 'slc', 'siu'])
-
-    # Loop through all the blueprint files we find that are referenced in gameClasses.json
-    for blueprint_file in sourcedir.joinpath('bp').glob('*.json'):
-        cls_name = blueprint_file.stem + '_C'
-        if not is_class_used(cls_name):
-            continue
-
-        blueprints = load_json_file(path=blueprint_file, quiet=True)
-
-        def bp_may_have_strings(bp):
-            return ((otype := bp.get('Type')) == cls_name or otype == 'TextRenderComponent') and 'Properties' in bp
-
-        # Loop through array of blueprints from json file that may have strings
-        for bp in (bp for bp in blueprints if bp_may_have_strings(bp)):
-
-            # Loop through game classes for that include the base class name (handles something like 'Coin:Chest_C')
-            for gc in (c for c in game_classes if cls_name in c):
-                props = bp['Properties']
-
-                if props.get('Text'):
-                    optKeyLoc(class_name=gc, k='friendly', v=props['Text'].get('SourceString'))
-                    optKeyLoc(class_name=gc, k='friendly_key', v=props['Text'].get('Key'))
-
-                if props.get('UpgradeName'):
-                    optKeyLoc(class_name=gc, k='friendly', v=props['UpgradeName'].get('SourceString'))
-                    optKeyLoc(class_name=gc, k='friendly_key', v=props['UpgradeName'].get('Key'))
-
-                if props.get('UpgradeDescription'):
-                    optKeyLoc(class_name=gc, k='description', v=props['UpgradeDescription'].get('SourceString'))
-                    optKeyLoc(class_name=gc, k='description_key', v=props['UpgradeDescription'].get('Key'))
-
-    print("Classes: ", len(classes), "Props: ", propcount, "Matches: ", propmatches)
-    save_json_file(data=game_classes, path=datadir.joinpath('gameClasses.json'))
-
-
-# Read custom-loc.json and merge it into gameClasses.json and write it out
-# Then collect all the keys found the various frontend data files and
-# come up with a stripped down version of the loc files with the keys for
-# all of them
-def export_loc_files(game: str, datadir: Path, sourcedir: Path) -> None:  # noqa: C901 - disable complexity warning
-    # Merge custom-loc.json into gameClasses.json
-    print('Merging custom-loc.json into gameClasses.json...')
-    classes = load_json_file(path=datadir.joinpath('gameClasses.json'))
-    customLoc = load_json_file(path=datadir.joinpath('custom-loc.json'))
-    for c, d in customLoc.items():
-        classes[c] = classes[c] | d
-    save_json_file(data=classes, path=datadir.joinpath('gameClasses.json'))
-
-    # Make a list of the keys that are referenced in the various files
-    keys = set()
-    key_files = [
-        'gameClasses.json',
-        'layerConfigs.json',
-        'custom-loc.json',
-        f'markers.{game}.json',
-        f'custom-markers.{game}.json',
-    ]
-    print('Reading files to determine loc keys required...')
-    for fn in key_files:
-        path = datadir.joinpath(fn)
-        if path.exists():
-            data = load_json_file(path=path)
-            for entry in data if type(data) is list else data.values():
-                for k in ['friendly_key', 'name_key', 'description_key', 'key']:
-                    if entry.get(k):
-                        keys.add(entry[k])
-
-    print(f'Found {len(keys)} keys')
-
-    print('Writing loc files to data directory')
-
-    # Loop through all languages present
-    for loc in sourcedir.joinpath('loc').glob('*/'):
-        loc_file = datadir.joinpath(f'loc/locstr-{loc.name}.json')
-
-        # Read existing keys if there are any
-        usedlocstr = load_json_file(path=loc_file) if loc_file.exists() else {}
-
-        # Read the games data for this language and add key/string pairs that we're interested in
-        # Note: for some reason file is of the form { "": {"loc key": "loc string"...  } } so
-        # get the dictionary entry for key ''
-        locstr = load_json_file(path=sourcedir.joinpath(f'loc/{loc.name}/Game.json'))['']
-        for k in locstr.keys():
-            if k in keys:
-                usedlocstr[k] = locstr[k]
-
-        save_json_file(data=usedlocstr, path=loc_file)
-
-
-def export_markers(game: str, datadir: Path, sourcedir: Path) -> None:  # noqa: C901 - disable complexity warning
-    maps = {}  # dictionary from map name to json data list
-    area_mtx = {}  # Transform for each area map geometry
-    data = []  # Output marker data
-
-    pipes = {}
-    objects = {}
-
-    data_lookup = {}
-
-    # Phase 1: Read all the map json files in and build a look up table for references
-    # Also get any area/map file matrices (for streaming levels)
-    for area in config[game]['maps']:
-        # Store the map data
-        maps[area] = load_json_file(path=sourcedir.joinpath('levels', f"{area}.json"))
-
-        # Go through all objects in the map data and store lookups for later
-        for oidx, o in enumerate(maps[area]):
-            if not o.get('Outer') or not (p := o.get('Properties')):
-                continue
-            oname = o['Name']
-            otype = o['Type']
-
-            # For maps that are divided into multiple files, there may be a LevelTransform for entities in that
-            # file relative to the persistent world that is handled by the streaming system. To correct for this
-            # we construct a matrix from the Translation/Rotation members if they exit
-            if (a := p.get('WorldAsset', {}).get('AssetPathName')) and (t := p.get('LevelTransform')):
-                area_mtx[a.split('.').pop()] = (
-                    Matrix.Translation(getVec(t.get('Translation'))) @ getQuat(t.get('Rotation')).to_matrix().to_4x4()
-                )
-
-            if otype.startswith('Pipesystem') and 'Pipe' in p and ('OtherPipe' in p or 'otherPipeInOtherLevel' in p):
-                # p['Pipe']
-                #     ObjectName: StaticMeshComponent'DLC2_Complete:PersistentLevel.HealingStation13_44.Pipe'
-                #     ObjectPath: SupralandSIU/Content/FirstPersonBP/Maps/DLC2_Complete.58973
-                # p['OtherPipe']
-                #     ObjectName: PipesystemNew_C'DLC2_Complete:PersistentLevel.PipesystemNew10'
-                #     ObjectPath: SupralandSIU/Content/FirstPersonBP/Maps/DLC2_Complete.19652
-                # p['otherPipeInOtherLevel']
-                #     AssetPathName: /Game/FirstPersonBP/Maps/DLC2_Complete.DLC2_Complete
-                #     SubPathString: PersistentLevel.PipeToArea2
-                # {ComponentType/class}'{map}'
-                def getPipeObjectName(o):
-                    t = o['ObjectName'].split('.')
-                    r = t[-2] if t[-1] == "Pipe'" else t[-1]
-                    return r.replace("'", "")
-
-                a = ':'.join((area, getPipeObjectName(p['Pipe'])))
-                if t := p.get('otherPipeInOtherLevel'):
-                    b = ':'.join((t['AssetPathName'].split('.').pop(), t['SubPathString'].split('.').pop()))
-                else:
-                    b = ':'.join((area, getPipeObjectName(p['OtherPipe'])))
-                pipes[a] = b
-                print(f"{a} -> {b}")
-                # pipes [b] = a # links may be single-sided
-
-            objects[area + ':' + oname] = o
-
-    for area in maps:
-        for oidx, o in enumerate(maps[area]):
-            otype = o['Type']
-            oname = o['Name']
-
-            allowed_items = (
-                marker_types
-                and otype in marker_types
-                or any(otype.startswith(s) for s in starts_with)
-                or any(otype.endswith(s) for s in ends_with)
-            )
-            if not allowed_items:
-                continue
-
-            if not o.get('Outer') or not (p := o.get('Properties')):
-                continue
-
-            def getObject(p):
-                ref = objectRef(p)
-                return maps[ref[0]][ref[1]]
-
-            def get_matrix(o, matrix=Matrix.Identity(4)):
-                p = o.get('Properties', {})
-
-                if p.get('RelativeLocation'):
-                    matrix = (
-                        Matrix.LocRotScale(
-                            getVec(p.get('RelativeLocation')),
-                            getRot(p.get('RelativeRotation')),
-                            getVec(p.get('RelativeScale3D'), 1),
-                        )
-                        @ matrix
-                    )
-
-                for prop in ['RootObject', 'RootComponent', 'DefaultSceneRoot', 'AttachParent']:
-                    if p.get(prop):
-                        return get_matrix(getObject(p[prop]), matrix)
-                return matrix
-
-            matrix = get_matrix(o)
-            if area in area_mtx:
-                matrix = area_mtx[area] @ matrix
-
-            # some MetalBall_C are Anvils, do the replacement
-            if (
-                o['Type'] == 'MetalBall_C'
-                and o.get('Properties', {}).get('Mesh?', {}).get('ObjectName') == "StaticMesh'Anvil'"
-            ):
-                o['Type'] = 'Anvil_C'
-
-            # some RingRusty_C are pickaxes, cannot be determined by meshes
-            if game == 'sl' and o['Name'].startswith('RingRusty'):
-                for i in range(10, 16 + 1):
-                    if o['Name'] == 'RingRusty' + str(i):
-                        o['Type'] = '_Pickaxe_C'
-                        # special type, an item
-
-            data.append({'name': o['Name'], 'type': o['Type'], 'area': area})
-            data_lookup[':'.join((area, o['Name']))] = data[-1]
-
-            t = matrix.to_translation()
-            data[-1].update({'lat': t.y, 'lng': t.x, 'alt': t.z})
-
-            p = o.get('Properties', {})
-
-            for key in properties:
-                optKey(data[-1], camel_to_snake(key), p.get(key))
-
-            def class_from_objectname(properties: dict, prop: str):
-                c = properties.get(prop, {}).get('ObjectName')
-                return c.split("'")[1] if c else None
-
-            optKey(data[-1], 'spawns', class_from_objectname(p, 'Spawnthing'))
-            optKey(data[-1], 'spawns', class_from_objectname(p, 'Class'))
-            optKey(data[-1], 'other_pipe', pipes.get(':'.join((area, o['Name']))))
-            optKey(data[-1], 'custom_color', optColor(p.get('CustomColor')))
-
-            if o['Type'] in ('Jumppad_C'):
-                optKey(data[-1], 'velocity', (v := p.get('Velocity')) and getXYZ(getVec(v)))
-                d = Vector((matrix[0][2], matrix[1][2], matrix[2][2]))
-                d.normalize()
-                data[-1].update({'direction': getXYZ(d)})
-                data[-1].update({'target': getXYZ(Vector((0, 0, 0)))})
-
-    calc_pads(data)
-    calc_pipes(data)
-
-    # Merge in custom and legacy data, clean the properties and remove ones we don't need
-    cleanup_objects(game=game, datadir=datadir, sourcedir=sourcedir, data_lookup=data_lookup, data=data)
-
-    print('collected %d markers' % (len(data)))
-    save_json_file(data=data, path=datadir.joinpath(f'markers.{game}.json'))
-
-
-def calc_pipes(data):
-    # I could not find hierarchy connection between pipe caps and pipe systems
-    # so I decided to search for the nearest pipe cap
-    # It should work fine most of the time
-    def allowed_points(o):
-        return o['type'] in ('PipeCap_C')
-
-    points = [(o['lng'], o['lat'], o['alt']) for o in data if allowed_points(o)]
-    data_indices = [i for i, o in enumerate(data) if allowed_points(o)]
-    print('collected', len(points), 'pipe caps, calculating links...')
-
-    if not points:
-        return
-
-    tree = KDTree(points)
-
-    lookup = {}
-    cap_indices = {}
-
-    # update pipe system, find nearest caps
-    for i, o in enumerate(data):
-        if not o['type'].startswith('Pipesystem'):
-            continue
-        x, y, z = o['lng'], o['lat'], o['alt']
-
-        query_point = [x, y, z]
-        _, indices = tree.query([query_point], k=3)
-        indices = indices[0]
-        j = data_indices[indices[0]]
-        p = data[j]
-        dist = (Vector((x, y, z)) - Vector((p['lng'], p['lat'], p['alt']))).length
-        if dist <= 1500:
-            nearest_cap = p['area'] + ':' + p['name']
-            cap_indices[nearest_cap] = j
-            data[i].update({'nearest_cap': nearest_cap})
-            a = o['area'] + ':' + o['name']
-            lookup[a] = o
-
-    ''' just add nearest cap for now ^. handle the rest in frontend
-    # update caps with cross-references
-    # not all pipes have caps, unfortunately
-    # some only have level geometry that's not in the classes
-    for i, o in enumerate(data):
-        if o['type'] not in ('PipesystemNew_C', 'PipesystemNewDLC_C'):
-            continue
-        # get nearest cap, get other pipe, update other pipe's cap
-        if 'nearest_cap' in o and 'other_pipe' in o:
-            if nearest_cap := o.get('nearest_cap'):
-                if other_pipe := o.get('other_pipe'):
-                    if p:=lookup.get(other_pipe):
-                        if other_cap := p.get('nearest_cap'):
-                            if j := cap_indices.get(nearest_cap):
-                                data[j].update({'other_cap': other_cap})
-    '''
-
-
-def calc_pads(data):  # noqa: C901 - disable complexity warning
-    # calculates target altitude from the jump pad's velocity data
-    # builds 3d terrain from selected points (uses jump pad locations by default)
-    # traces parabolic path and find z of an intersection with a plane defined by 3 closest points
-
-    def allowed_points(o):
-        # return True   # not recommended, bugs with coins, etc.
-        return o['type'] not in (
-            'Coin_C',
-            'BP_TriggerVolume_C',
-            'Button_C',
-            'Door_C',
-            'Chest_C',
-        )  # testing
-
-    # def allowed_points(o):
-    #     return o['type'] in ('Jumppad_C') # jump pads only
-
-    points = [(o['lng'], o['lat'], o['alt']) for o in data if allowed_points(o)]
-    # data_indices = [i for i, o in enumerate(data) if allowed_points(o)]
-
-    print('collected', len(points), 'terrain points, calculating targets...')
-
-    if not points:
-        return
-
-    tree = KDTree(points)
-    jumppads = []
-    matches = {}
-
-    for i, o in enumerate(data):
-        if o['type'] != 'Jumppad_C':
-            continue
-        jumppads.append(o)
-        matches[':'.join((o['area'], o['name']))] = {'obj': o, 'targets': []}
-
-        x, y, z = o['lng'], o['lat'], o['alt']
-        k = o.get('relative_velocity', 1000)
-        v = o.get('direction', {'x': 0, 'y': 0, 'z': 0})
-
-        vx = v['x'] * k
-        vy = v['y'] * k
-        vz = v['z'] * k
-
-        if (v := o.get('velocity')) and o.get('allow_stomp'):
-            vx = v['x']
-            vy = v['y']
-            vz = v['z']
-
-        dt = 0.01
-        g = 9.8
-        m = 95
-        h = 0
-        t = 0
-        last_z = z
-        s = Vector((x, y, z))
-        while t < 20:
-            vz -= g * m * dt
-            x += vx * dt
-            y += vy * dt
-            z += vz * dt
-            t += dt
-
-            query_point = [x, y, z]
-            _, indices = tree.query([query_point], k=3)
-            indices = indices[0]
-            triangle = [points[j] for j in indices]
-            h = get_z(x, y, triangle)
-
-            dist = (Vector((x, y, z)) - s).length
-
-            # print([round(v, 2) for v in [x, y, z]], 'h', round(h, 2), 'nearest triangle', [ data[data_indices[j]]['name']+':'+str([round(x, 2)for x in points[j]]) for j in indices])
-            if dist > 250 and last_z > z and h > z:  # only check on decline
-                break
-
-            last_z = z
-
-        # print('pad', o['name'], 'velocity', vx, vy, vz, 'target', x, y, z)
-        data[i].update({'target': {'x': x, 'y': y, 'z': z}})
-
-    # Now we try to find pairs of jumppads. For each jump pad, find all the other jumppads close to the target
-    # If two jumppads are both in each other's lists then we can consider merging
-
-    def getdir(o):
-        v = o.get('direction', {'x': 0, 'y': 0, 'z': 0})
-        vx = -v['x']
-        vy = -v['y']
-        vz = v['z']
-
-        if (v := o.get('velocity')) and o.get('allow_stomp'):
-            vx = v['x']
-            vy = v['y']
-            vz = v['z']
-        return Vector((vx, vy, vz)).normalized()
-
-    for o in jumppads:
-        if abs(getdir(o).z) > 0.99:
-            continue
-
-        ov = Vector((o['lng'], o['lat'], 0))
-        otv = Vector((o['target']['x'], o['target']['y'], 0))
-        do = (otv - ov).normalized()
-
-        tdist = (otv - ov).length  # Distance between object and its target
-
-        for tj in jumppads:
-            if tj is o:
-                continue
-            if abs(getdir(tj).z) > 0.99:
-                continue
-
-            jv = Vector((tj['lng'], tj['lat'], 0))
-            jtv = Vector((tj['target']['x'], tj['target']['y'], 0))
-            dt = (jtv - jv).normalized()
-
-            no2j = (jv - ov).normalized()
-            dist = (otv - jv).length
-            # distance between target and prospective match
-
-            # Distance threshold is compared to the distance between the pads
-            # Check that the pads are pointing in opposite directions
-            # And that the pads are facing each other (rather than opposite directions)
-            if dist < 0.3 * tdist and do.dot(dt) < -0.98 and no2j.dot(do) > 0.97 and no2j.dot(dt) < -0.97:
-                # print(o['name'], tj['name'], do.dot(dt), no2j.dot(do), no2j.dot(-dt), round(dist/tdist, 2))
-                alt = ':'.join((o['area'], o['name']))
-                matches[alt]['targets'].append(tj)
-
-    plist = {}
-    for m in matches.values():
-        o = m['obj']
-        for t in m['targets']:
-            if t is o:
-                continue
-            alt = ':'.join((t['area'], t['name']))
-            for tt in matches[alt]['targets']:
-                if tt is o:
-                    if not plist.get(o['name']) and not plist.get(t['name']):
-                        plist[o['name']] = t['name']
-                        o['target']['x'] = t['lng']
-                        o['target']['y'] = t['lat']
-                        o['target']['z'] = t['alt']
-                        o['other_pad'] = alt
-                        o['twoway'] = 1
-                        t['target']['x'] = o['lng']
-                        t['target']['y'] = o['lat']
-                        t['target']['z'] = o['alt']
-                        t['twoway'] = 2
-                        t['other_pad'] = ':'.join((o['area'], o['name']))
-
-
-def get_z(x, y, triangle):
-    v1, v2, v3 = triangle
-    denominator = (v2[1] - v3[1]) * (v1[0] - v3[0]) + (v3[0] - v2[0]) * (v1[1] - v3[1])
-    if denominator == 0:
-        return v1[2]
-    alpha = ((v2[1] - v3[1]) * (x - v3[0]) + (v3[0] - v2[0]) * (y - v3[1])) / denominator
-    beta = ((v3[1] - v1[1]) * (x - v3[0]) + (v1[0] - v3[0]) * (y - v3[1])) / denominator
-    gamma = 1 - alpha - beta
-    if alpha >= 0 and beta >= 0 and gamma >= 0:
-        z = alpha * v1[2] + beta * v2[2] + gamma * v3[2]
-    else:
-        alpha = max(0, min(1, alpha))
-        beta = max(0, min(1, beta))
-        gamma = 1 - alpha - beta
-        z = alpha * v1[2] + beta * v2[2] + gamma * v3[2]
-    return z
-
-
-# Variant names for the colours of keys, keycards, flowers, seeds...
-colors = {
-    0: 'white',  # FFFFFF
-    1: 'yellow',  # FFFF66
-    2: 'red',  # FF0000
-    3: 'blue',  # 1E90FF
-    4: 'purple',  # 9933ff
-    5: 'green',  # 4DFF00
-    6: 'orange',  # ff9900
-}
-
-# Variant names for minecraft bricks
-brick_types = {
-    0: 'stone',
-    1: 'obsidian',
-    2: 'metal',
-    3: 'diamond',
-    4: 'gold',
-}
-
-price_types = [
-    "coin",
-    "coal",
-    "iron",
-    "diamond",
-    "supranium",
-    "scrap",
-    "bone",
-]
-
-# Number of coins given by classes if not explicit
-# Coin pots provide 1 if the flag is true
-coin_defaults = {
-    'Coin_C': 1,
-    'CoinBig_C': 10,
-    'LotsOfCoins1_C': 1,
-    'LotsOfCoins5_C': 5,
-    'LotsOfCoins10_C': 10,
-    'LotsOfCoins15_C': 15,
-    'LotsOfCoins30_C': 30,
-    'LotsOfCoins50_C': 50,
-    'LotsofCoins200_C': 200,  # Note lower case 'of'
-    'PhysicalCoin_C': 1,
-    '5Cent_C': 5,
-    'RealCoinPickup_C': 1,
-    'Inventory_Coin_C': 1,
-    'Inventory_Coin3_C': 3,
-    'Inventory_Coin5_C': 5,
-    'Inventory_Coin10_C': 10,
-    'Gumball_Machine_C': 0,
-}
-
-# Which properties we allow to be exported for each instance
-exported_properties = [
-    'name',
-    'type',
-    'area',
-    'lat',
-    'lng',
-    'alt',  # all instances have these
-    'spawns',  # generates some other class (chest or quest)
-    'coins',  # generates coins when taken (coin chest, pots, bricks and various coin types)
-    'cost',
-    'price_type',  # cost when purchased (shops, quests...) and units (defaults to coins)
-    'icon',  # explicit icon override
-    'variant',  # variant allows change of marker
-    'friendly',  # Friendly name for the marker
-    'other_pipe',  # For pipes we store the pipe at the other end
-    'nearest_cap',  # Nearest pipecap we are connected to if there is one
-    'other_pad',  # For two way pads we store the pad at the other end
-    'linetype',  # Trigger, pad, pipe, target
-    'twoway',  # Pipe or pad is two way if True
-    'notsaved',  # Pipe or pad is not saved
-    'target',  # where to draw line to for pipes and pads
-    'targets',  # array of dictionaries 'type' and target position
-    'old_coins',  # For _CoinStack_C's a dictionary of old coin name to value (for save game handling)
-    'scrapamount',  # For Scrap_C gives amount of scrap
-    # 'image',
-    'yt_video',
-    'yt_start',
-    'yt_end',  # data pulled from matched legacy data
-]
-
-
-# The purpose of this code is to walk through all the objects we've gathered and prepare them for
-# display by the map.
-#
-# classes is a set of all 'type' names
-# lookup is a dictionary from area:name to objects
-# data is an array of the same objects
-# each object is a dictionary of k, v pairs
-def cleanup_objects(  # noqa: C901 - disable complexity warning
-    game: str, datadir: Path, sourcedir: Path, data_lookup: dict, data: list[dict]
-) -> None:
-    def get_xyz(o):
-        return dict(x=o['lng'], y=o['lat'], z=o['alt'])
-
-    def get_nc_xyz(o):
-        return get_xyz(data_lookup[o['nearest_cap']]) if 'nearest_cap' in o else get_xyz(o)
-
-    # Read the set of pads and pipes we found save data for
-    savedpadpipes = read_savedpadpipes(game=game, sourcedir=sourcedir)
-
-    # Filter out classes that won't get used
-    # Note: consider checking for 'dev' layers as well as undefined
-    game_classes = load_json_file(path=datadir.joinpath('gameClasses.json'))
-
-    def is_class_used(item: dict) -> bool:
-        return (
-            (gc := game_classes.get(item['type']))
-            and game in gc.get('games', ['sl', 'slc', 'siu'])
-            and (gc.get('layer') or gc.get('nospoiler'))
-        )
-
-    for item in data[:]:
-        if not is_class_used(item):
-            data.remove(item)
-
-    # Walk the remaining instances and fix up entries
-    # We loop over a copy to allow us to remove entries we don't want
-    for o in data[:]:
-        alt = ':'.join((o['area'], o['name']))
-
-        # Merge the various gold properties into one (we'll remove the properties later)
-        # and deal with defaults for all coin or coin spawning classes
-        # also takes care of DestroyablePots_C
-        coins = o.get('coins') or o.get('coins_in_gold') or o.get('value') or (1 if o.get('contains_coin') else None)
-        if coins is not None:
-            o['coins'] = coins
-        else:
-            if o['type'] in coin_defaults:
-                o['coins'] = coin_defaults[o['type']]
-            elif o.get('spawns') in coin_defaults:
-                o['coins'] = coin_defaults[o['spawns']]
-
-        # There is at least one chest that doesn't have spawn data but actually spawns Health+1 (SL Chest31_9005)
-        if o['type'] == 'Chest_C' and o.get('spawns') is None:
-            o['spawns'] = 'BP_PurchaseHealth+1_C'
-
-        # If this is the stolen coins chest then set coins to 'some' it will remove the spawns property later
-        if o.get('spawns') == 'StolenCoins_C':
-            o['coins'] = 'varies'
-
-        # Minecraft bricks may contain coins, but if they are not specified gold ones default to 3
-        # variant is set to the brick type
-        if o['type'] == 'MinecraftBrick_C':
-            if game == 'siu':
-                if type(bt := o.get('brick_type')) is str:
-                    o['variant'] = bt.rsplit('::')[-1].lower()
-                else:
-                    o['variant'] = brick_types[4 if not bt and o.get('coins') else 0]
-                if o['variant'] == 'gold' and not o.get('coins'):
-                    o['coins'] = 3
-                if o.get('coins') is not None and o['variant'] != 'gold':
-                    print(f'Non gold coin containing MinecraftBrick_C:{o["variant"]}:{o["coins"]}')
-
-        # Player icon changes colour based on game
-        if o['type'] == 'PlayerStart':
-            o['variant'] = 'blue' if game == 'siu' else 'red'
-
-        # Deal with Waldo's in SLC
-        if o['type'] == 'RedGuy_C' and o['name'].startswith('Waldo'):
-            o['type'] = 'Waldo:' + o['type']
-
-        # Allow things with colors can have variants (minecraft bricks have type and colour)
-        if o.get('variant') is None:
-            color = o.get('color') or o.get('original_color')
-            if color and type(color) is int and color >= 0 and color < len(colors):
-                o['variant'] = colors[color]
-            elif color and type(color) is str and (color := color.split('::')[-1]):
-                o['variant'] = color.lower()
-
-        if (pt := o.get('price_type')) is not None and type(pt) is str:
-            o['price_type'] = price_types.index(pt.split('::')[-1].lower())
-
-        # Anything that has coins gets a subclass (chests, minecraft bricks, destroyable pots...)
-        # Anything that provides coins and spawns we clear the spawns field (chests can't do both)
-        # There's also a chest that has coins in it but also has a "spawns" (SIU ChestAreaEnd_78)
-        if o.get('coins') is not None:
-            if o['type'] not in coin_defaults:  # don't add subclass to coins
-                o['type'] = 'Coin:' + o['type']
-            if o.get('spawns') is not None:
-                del o['spawns']
-
-        # Create line data
-
-        if o.get('other_pipe'):
-            o['linetype'] = 'pipe'
-            if o.get('nearest_cap'):
-                # We used to put this in startpos but we just move the pipe now
-                nc = data_lookup[o['nearest_cap']]
-                o['lat'], o['lng'], o['alt'] = nc['lat'], nc['lng'], nc['alt']
-            if o['other_pipe'] == alt:
-                del o['linetype']
-                del o['other_pipe']  # Some pipes point at themselves (basically in only)
-            else:
-                opo = data_lookup[o['other_pipe']]
-                o['target'] = get_nc_xyz(opo)
-                if alt == opo.get('other_pipe'):
-                    o['twoway'] = 1
-                    opo['twoway'] = 2
-        elif o['type'] == 'Jumppad_C':
-            o['linetype'] = 'jumppad'
-            if (o.get('allow_stomp')
-                    or o.get('disable_movement_in_air') == False):   # fmt: skip # noqa: E712 default for DisableMovementInAir is True
-                o['variant'] = 'blue'
-            else:
-                o['variant'] = 'red'
-
-        # Mark pads/pipes for which we haven't identified save information
-        if o['type'] == 'Jumppad_C' and alt not in savedpadpipes['pads']:
-            o['notsaved'] = True
-        if (
-            'Pipesystem' in o['type']
-            and alt not in savedpadpipes['pipes']
-            and ((nc := o.get('nearest_cap')) and nc not in savedpadpipes['pipes'] or not nc)
-        ):
-            o['notsaved'] = True
-
-    # Convert piles of non-rotating coins to stack markers
-    create_coinstacks(data_lookup, data)
-
-    # Strip properties we don't need to export
-    for o in data:
-        for prop in list(o.keys()):
-            if prop not in exported_properties:
-                del o[prop]
-
-
-# Goes through all the non-rotating coins and looks for groups of more than 3 to combine into coinstacks
-# Adds the new stack objects to our collection and removes the original coins
-def create_coinstacks(data_lookup, data) -> None:
-
-    threshold = 400
-
-    # Positions of all non-rotating coins and their corresponding indices in data
-    points = []
-    dataidx = []
-    for idx, o in enumerate(data):
-        if o['type'] in ['Coin_C', 'CoinBig_C'] and o.get('is_doesnt_rotate'):
-            points.append([o['lng'], o['lat'], o['alt']])
-            dataidx.append(idx)
-
-    if len(points) == 0:
-        return
-
-    # Creates a graph of connected elements that are within the distance threshold ie sets of
-    # points that all fit within a sphere of max radius 'threshold'
-    graph = weights.DistanceBand.from_array(points, threshold=threshold, silence_warnings=True).to_networkx()
-
-    # Create an array of new stack objects to add and a list of old coin indices to remove
-    stackId = 0
-    stacks = []
-    delobj = []
-    for cc in nx.connected_components(graph):
-        # Note inc should be adfter the condition however there is custom
-        # data depending on the number now and it only matters they are unique
-        stackId += 1
-        if len(cc) > 3:
-            coins = 0
-            old_coins = {}
-            for idx in cc:
-                o = data[dataidx[idx]]
-                coins += o['coins']
-                old_coins[o['name']] = o['coins']
-                area = o['area']
-                delobj.append(data[dataidx[idx]])
-            c = np.r_[points][list(cc)].mean(axis=0)
-
-            stacks.append(
-                {
-                    'name': f'CoinStack{stackId}',
-                    'type': '_CoinStack_C',
-                    'area': area,
-                    'lat': c[1],
-                    'lng': c[0],
-                    'alt': c[2],
-                    'coins': str(coins),
-                    'old_coins': old_coins,
-                }
-            )
-
-    # Delete the old coins and add the new ones - we don't bother clearing out data_lookup
-    delcount = len(delobj)
-    for o in delobj:
-        data.remove(o)
-
-    for stack in stacks:
-        data.append(stack)
-        data_lookup[':'.join((stack['area'], stack['name']))] = stack
-
-    print(f'Replaced {delcount} coins with {len(stacks)} stacks')
-
-
-# Somewhat hacky code to convert class name to a more friendly name
-def friendly_name(cls: str):
-    n = re.sub(r'(?:_C$)|(?:^BP_)|(?:Purchase)|(?:Buy)|_|DLC2|DLC|New', '', cls)
-    n = n.replace('pct', '%')
-    n = re.sub(r'x([0-9])', r'*\1', n)
-    n = re.sub(r'([A-Z+-]+)', r'.\1', n)
-    n = re.sub(r'([a-z])([0-9])', r'\1.\2', n)
-    n = re.sub(r'([0-9])([a-z]{2, 20})', r'\1.\2', n)
-    n = n.replace('Smashdown', 'Stomp')
-    n = n.replace('*', 'x')
-    n = n.rstrip('.').lstrip('.')
-    n = n.replace('.', ' ')
-    return n
-
-
-# This file is generated using hard coded information and info from a save
-# file. It is generated by dumpsavefile.js
-def read_savedpadpipes(game: str, sourcedir: Path) -> dict[str, list]:
-    path = sourcedir.joinpath(f'savedpadpipes.{game}.json')
-
-    # Open and read the whole js file if it exists
-    if path.exists():
-        print(f'Reading {path}...')
-        return load_json_file(path=path)
-
-    print(f'Warning: {path} not found, skipping')
-    return {'pipes': [], 'pads': []}
+from parserlib.localisation import export_class_loc, export_loc_files
+from parserlib.slmarkers import export_markers
+from parserlib.swmarkers import export_sw_markers
+from parserlib.swpreproc import preproc_levels
+from parserlib.swversion import update_swversion_info
 
 
 def main() -> None:
